Resolve #386: Automatically test serialization compatibility
Tests serialization compatibility for the various Input/Output
variants: an instance of `TestData` was serialized to a file
(one for each IO variant), which is deserialized and compared
to a new `TestData` instance (the expected value). Because Unsafe
uses native byte ordering, files for Unsafe based variants are
specific to the endianness used when they were created (le for
little-endian).

The TestData class contains fields for each default serializer
(respectively the related type) and some arbitrary data structure
(`Person`) that shall mimick some real world data (perhaps not
really needed, just to have a class that's not too trivial).

If a new default serializer is added, the `TestData` should be
altered, and data files should/must be recreated (just delete
the existing test data and run the test to create new files).

Implementation notes:

* Needed to override KryoTestCase.assertEquals because the standard
  junit assert equals does not handle object graphs accordingly.
  commons EqualsBuilder.reflectionEquals is also not sufficient, therefore
  I added a reflectionEquals I used in kryo-serializers and memcached-
  session-manager and in other projects (so it's fairly good tested).
  Also tried unitils' `ReflectionAssert.assertReflectionEquals`, but
  this failed for StringBuilder/StringBuffer.