Allow to override type by name lookup in DefaultClassResolver.
This shall support (de)serialization of cglib proxies.
A cglib proxy is typically serialized with a class name like e.g.
  FOO_org.example.MyClass$$EnhancerByCGLIB$$580cffdd
which, during deserialization, is the input to DefaultClassResolver.readName.

Because readName invokes Class.forName such proxies could not be
deserialized (Class.forName obviously would fail), this change shall
make it possible to resolve a valid type for such class names (a valid
type is a type for that a Registration exists).

For a real use case see CGLibProxySerializer (and Test) in kryo-serializers.