Fix #189 Don't embed minlog Logger in jar
The felix bundle plugin added the minlog classes due to the
Export-Package and Embed-Dependency declarations with value
com.esotericsoftware*.
The Embed-Dependency decl is now removed (seems duplicate with shading),
the Export-Package explicitely sets the kryo and reflectasm
subpackages. With this the minlog classes are no longer contained in
the jar.