Add automated compatibility check (with previous version)
Uses http://mojo.codehaus.org/clirr-maven-plugin/plugin-info.html
For configuration see
http://mojo.codehaus.org/clirr-maven-plugin/check-mojo.html

The compat check is bound to compile, on my laptop it adds ~0.8sec
on top and is woth it IMHO. Alternatively it could be bound to
test, or just be run manually via `mvn clirr:check`.

A report can be generated using `mvn clirr:clirr`.

Closes #150