Change default artifact *not* to shade reflectasm, use OSGi'ed reflectasm 1.3.0
There are not two artifacts: kryo and kryo-shaded, where kryo
has normal dependencies on 3rd party artifacts (and on the normal
reflectasm as well that depends on asm), and kryo-shaded does
*not* depend on reflectasm but has reflectasm classes and relocated
asm classes included (in the kryo-shaded jar) to prevent conflicts
with other versions of asm that might be in the classpath.

The kryo artifact is now (because reflectasm classes are no longer
contained in the jar) binary incompatible with the previous version,
while kryo-shaded is compatible with the previous version.

The dependency on reflectasm is updated to 1.3.0 which is now OSGi
conform, so the standard kryo artifact should work in an OSGi environment.

This change was discussed on the mailing list:
https://groups.google.com/forum/#!topic/kryo-users/CYgpDylwvf4