Change groupId to com.esotericsoftware.
This has been discussed some times previously and now it seems to be
ok doing this. As a consequence the minor version will be changed
in another commit.

The clirr compatibility check is now temporarily disabled because
it would compare with kryo 1.04. With the next release it should
be enabled again.