Disable doclint to fix build/release error with java 8.
Building the javadoc jar failed with java8 due to invalid javadoc,
which was detected by doclint.

From http://stackoverflow.com/questions/15886209/maven-is-not-working-in-java-8-when-javadoc-tags-are-incomplete:

DocLint is a new feature in Java 8, which is summarized as:
  Provide a means to detect errors in Javadoc comments early in
  the development cycle and in a way that is easily linked back
  to the source code.

This is enabled by default, and will run a whole lot of checks
before generating Javadocs.