Exclude java8 tests for jdks < 1.8
After having built tests with jdk 8 with

   $ mvn -f pom-main.xml test-compile // or 'test' or 'verify'

tests can be run with an older jdk with

    $ mvn -f pom-main.xml surefire:test

so that classes/tests are not compiled again but the compiled tests
are just executed. The newly introduced build profile excludes all tests
in the `java8` package for jdks older than 1.8.
The profile is added to the parent pom (instead of pom-main.xml) without
any hard reason, but just to keep profiles at one place.

Refs #360