Move Generics{,Resolver} to serializers, reduce api, go back to 3.1.0-SNAPSHOT
This is related to #342 (last change to Generics), and also related
to #343 (reduce api)

Because Generics(Resolver) related methods are not considered to be
part of the public api, no user should be affected. The removed methods
and moved classes are added to the clirr ignore list.

The project version is changed back to 3.1.0-SNAPSHOT, because it had
been accidentally incremented to 4.0.0 (now it's clear that the major
version shall only be increased when serialization compatibility is
broken).