Adds support for java 8 java.time.* classes
The serializers are added as default serializers for java 8
(as determined via the `java.version` system property that should
be available in all jvms.

The following classes are supported:

* Duration
* Instant
* LocalDate
* LocalTime
* LocalDateTime
* ZoneOffset
* ZoneId
* OffsetTime
* OffsetDateTime
* ZonedDateTime
* Year
* YearMonth
* MonthDay
* Period

`ZoneRegion` is the second implementation of `ZoneId` besides `ZoneOffset`
but because it's not public it must be registered via e.g.
`kryo.register(ZoneId.of("UTC+01:00").getClass());`.

The serialization of these classes is aligned to the serialization
within these classes itself (as used by `java.time.Ser`).

Also updates the java 8 test data (and files) to have fields for the new
default serializers.

Resolves #358