Get rid of java8 package, move OptionalSerializers to serializers package
To use a dedicated package for the java version seems too artificial. One
motivation was to be able to generally exclude tests via a maven profile,
however this can also easily be done with junits Assume to abort/skip
a test if the java version does not match.

JavaVersion is also moved to utils now and technically public, but marked
as part of the internal api and may therefore change.