Fix test: Activate junit4 for proper handling of Assume
Tests inheriting from KryoTestCase are run by junit 3 runner
which does not support Assume to skip tests (which was introduced
in 4.4).

Changing KryoTestCase to run with junit 4 would need changes to
various tests: assert* would have to be imported,  @Test would
have to be added and @Before would have to be used.

Therefore the change is kept to a minimum, in the test class that
needs junit 4 features.

For upgrading tests to junit 4 I've created
https://github.com/EsotericSoftware/kryo/issues/397