Introduce FieldSerializerConfig to encapsulate config in Kryo
The `Kryo` class is more and more used for "global" configuration of
serializers like `FieldSerializer` etc (here's another PR: #365).
The semantics of such settings and especially for which serializer
it's applied can often only be clarified via documentation.

This change introduces a `FieldSerializerConfig` class that encapsulates
configuration for `FieldSerializer`. It allows to configure settings
for all `FieldSerializer` instances created via `Kryo.getFieldSerializerConfig`.
On each `FieldSerializer` the config can be changed/overridden, this is
done directly on the serializer (instead e.g. requiring to invoke
`fieldSerializer.getConfig.setSomeThing`) to stay binary compatible and
because forcing to use the config here does not seem to be of much value.
The `FieldSerializer` internally uses the config object so that any further
settings are directly added to `FieldSerializerConfig` and there's only
a single place to keep config settings.

`Kryo.asmEnabled` might be used by different serializers, but because until
now it was only used by `FieldSerializer` this setting is also kept in
`FieldSerializerConfig`. `Kryo.set/getAsmEnabled` is marked as deprecated
with a hint to use `getFieldSerializerConfig` instead - this is done to
stay binary compatible.

`Kryo.setCopyTransient` is removed because it was introduced after the
latest release 3.0.3 and therefore not yet a published/released. Now
`Kryo.getFieldSerializerConfig.setCopyTransient` has to be used.

If this is considered to be merged then #365 should also follow this
approach.