Add test for java 8 closure serialization
For this test the compiler source/target level for tests must be set to 1.8.
This is done via the "java8" profile which is activated if built with
java 8+. Java8*Tests are excluded by default from tests compilation, this
profile removes this exclude. Because overriding the compiler plugin in the
profile did not really work build properties are used.

To deactivate the java8 profile (e.g. to compile tests for jdks < 1.8)
mvn can be invoked with `-P !java8` (or `-P \!java8` on linux).

Refs #299