Disable the optimization of generics serialization and provide an API for enabling/disabling it.
The existing optimization of generics serialization aims at reducing the size of the serialized representation. This optimization is enabled by default. But sometimes it may have a negative impact on performance. And sometimes it cannot handle complex cases of generics usage.

This patch changes the default behavior and disables this this optimization. To enable it, simply invoke `FieldSerializerConfig.setOptimizedGenerics(true)`.