Organized imports and formatted source.
Sorry for the large diff, but let's keep things tidy! This may result in poor formatting for a small amount of code. Such code can use `//` to create a newline that the formatter respects. Also @off and @on can be used to exclude code from being formatted (though rarely needed).

Eclipse formatter settings are (have always been) in Git.