Fix #529: Support serializing the Enum class object
When an `Enum.class` instance is serialized (by `ClassSerializer`), this
causes the creation of the registered default serializer (`EnumSerializer`)
with this call chain:
```
ClassSerializer.write
 -> Kryo.writeClass
  -> DefaultClassResolver.writeClass
   -> Kryo.getDefaultSerializer
    -> ReflectionSerializerFactory.makeSerializer(kryo, EnumSerializer, Enum.class)
```

This is done so that `Kryo.writeClass` can write the class name or the id
of the serializer registration.

Until now, the `EnumSerializer` failed instantiation for a type with `null`
enum constants. This however is the case for `Enum.class` itself.
The creation of `EnumSerializer` for `Enum.class` is now allowed, under the
assumption that this happens in the context of the `Enum.class` serialization,
and that `write`/`read` on the created `EnumSerializer` instance will never
be called. This assumption should be safe, since there should never be an instance
of the abstract `Enum` class.