Add versioned kryo artifact (currently kryo5)
This adds a completely self-contained kryo artifact that can be
used together with another major kryo version at the same time.

Therefore all dependencies are shaded to prevent potential conflicts.
The asm (shaded by reflectasm) that was placed into `com.esotericsoftware.asm`
is moved to `com.esotericsoftware.kryo5.asm` in order to prevent
conflicts.

The final artifact id is `kryo5`, similarly the Automatic-Module-Name (JPMS)
and the bundle name (OSGI).

The README should still be extended/adjusted to mention this artifact.

Refs #650