Remove osgi bundle stuff from versioned artifact
Because the bundle plugin printed several warnings, and the
osgi bundle related manifest entries were not correct (they were
missing kryo packages in `Export-Package`.

Because OSGI handles versioning and visibility, OSGI users
should want to use the unversioned artifact.

Therefore it's not worth it to attempt to somehow fix the OSGI
related build stuff for the versioned jar.