Issue #663 fix.
This addresses two issues in the CompatibleFieldSerializer.

1) The reported issue; the code now validates that the type read from
the bytecode is compatible with the type read from the serialized data
and skips it/throws (depending on chunk mode) if the type is not
compatible.

2) In implementing the above check, the unit tests began failing. It
turns out that the write method in CompatibleFieldSerializer had a field
cache corruption for primitive types due to autoboxing. This has been
mitigated by checking if the cached type is a primitive before
overwriting it with a value type. Given the primitive types have
dedicated CachedField implementations, this is perfectly safe and has no
binary compatibility concerns.