Fix mvn release for versioned artifact
The `mvn release:prepare` failed with
```
[ERROR] Failed to execute goal org.apache.maven.plugins:maven-shade-plugin:3.2.1:shade (default) on project kryo5: Error creating shaded jar: /path/to/kryo/target/kryo-5.0.0-RC5.jar (No such file or directory)
```
because release:prepare invoked clean when building the versioned (kryo5) module.

To fix this, for the maven release plugin the `preparationGoals`
are overridden, to only invoke `verify` (default is `clean verify`).

When running a release, `clean` of course has to be run manually at first.