Optimizations for `BigDecimal` default serializer.
Main purpose is to avoid inflating `BigDecimal` during reads,
i.e. avoiding creation of `BigInteger` unscaled value as if it
is passed to `BigDecimal` constructor it makes this object bigger,
i.e. instead of 40 bytes it would be 104 bytes. Additionally now
the throughput of serialization/deserialization is better for some
values.